Avoid race between proposeFuture and hasFailure
The provided hasFailure AtomicBoolean will be set by the callback
registered to proposeFuture, which will be executed concurrently with
RepairCompleteCallback without any guaranteed execution order. This can
lead to races when RCC tries to evaluate hasFailure, which hasn't been
set yet at this point. Discovered running CoordinatorMessagingTest.

I didn't want to modify RCC, so I've created a simple wrapping future
around the hasFailure callback that we return, instead of returning
proposeFuture directly. It should be the most non-invasive way to handle
this, but the code could use some cleanup when it comes to success/failure
handling.